+++ hit-core-api/src/main/java/gov/nist/hit/core/api/ValidationLogController.java
@@ -15,9 +15,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import gov.nist.auth.hit.core.domain.Account;
 import gov.nist.auth.hit.core.domain.ValidationLog;
 import gov.nist.hit.core.domain.ResponseMessage;
 import gov.nist.hit.core.domain.ResponseMessage.Type;
+import gov.nist.hit.core.service.AccountService;
 import gov.nist.hit.core.service.TestStepService;
 import gov.nist.hit.core.service.UserService;
 import gov.nist.hit.core.service.ValidationLogService;
@@ -39,38 +41,48 @@ public class ValidationLogController {
 	@Autowired
 	protected TestStepService testStepService;
 
+	@Autowired
+	private AccountService accountService;
+
 	private void checkPermission(Authentication auth) throws Exception {
 		String username = auth.getName();
 		if (username == null)
 			throw new NoUserFoundException("User could not be found");
-		if (!userService.hasGlobalAuthorities(username)) {
+		Account account = accountService.findByTheAccountsUsername(username);
+		if (account == null) {
+			throw new NoUserFoundException("You do not have the permission to perform this task");
+		}
+		if (!userService.hasGlobalAuthorities(username) && !userService.isAdminByEmail(account.getEmail())) {
 			throw new NoUserFoundException("You do not have the permission to perform this task");
 		}
 	}
 
-	@PreAuthorize("hasRole('admin')")
+	@PreAuthorize("hasRole('tester')")
 	@ApiOperation(value = "get all logs", nickname = "getAll")
 	@RequestMapping(method = RequestMethod.GET, produces = "application/json")
 	public List<ValidationLog> getAll(Authentication authentication, HttpServletRequest request,
 			HttpServletResponse response) throws Exception {
 		logger.info("Fetching all validation logs...");
+		checkPermission(authentication);
 		return validationLogService.findAll();
 	}
 
-	@PreAuthorize("hasRole('admin')")
+	@PreAuthorize("hasRole('tester')")
 	@RequestMapping(value = "/count", method = RequestMethod.GET, produces = "application/json")
 	public long countAll(Authentication authentication, HttpServletRequest request, HttpServletResponse response)
 			throws Exception {
 		logger.info("Fetching all validation logs count...");
+		checkPermission(authentication);
 		return validationLogService.countAll();
 	}
 
-	@PreAuthorize("hasRole('admin')")
+	@PreAuthorize("hasRole('tester')")
 	@ApiOperation(value = "delete log", nickname = "getAll")
 	@RequestMapping(value = "/{id}/delete", method = RequestMethod.POST, produces = "application/json")
 	public ResponseMessage deleteLog(Authentication authentication, HttpServletRequest request,
 			@PathVariable("id") Long id, HttpServletResponse response) throws Exception {
 		logger.info("deleting validation log with id=" + id + "...");
+		checkPermission(authentication);
 		validationLogService.delete(id);
 		return new ResponseMessage(Type.success, "Validation Log " + id + " deleted successfully", id + "", true);
 	}
